pyproject.toml
[project]
name = "discord-gradiusbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.5.2",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "ruff>=0.11.8",
]
